add employee and company files with their custom types
