+++ Employee.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace classes
+{
+    public class Employee
+    {
+        public string FirstName { get; set; }
+
+        public string LastName { get; set; }
+
+
+        public string Title { get; set; }
+
+        public string StartDate { get; set; }
+    }
+}
